backend/pyproject.toml
[project]
name = "jobs-rag"
version = "0.1.0"
description = "RAG agent for company information using LangChain and AI"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3",
    "langchain-google-genai>=2.0",
    "langchain-community>=0.3",
    "langchain-text-splitters>=0.3",
    "langsmith>=0.2",
    "chromadb>=0.5",
    "tavily-python>=0.5",
    "python-dotenv>=1.0",
    "langgraph>=0.2",
    "langchain-tavily>=0.2.16",
    "langchain-chroma>=1.1.0",
    "langchain-openai>=1.1.7",
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
